python-backend/requirements.txt
flask==2.3.3
pymongo==4.6.1
pyyaml==6.0.1
bcrypt==4.0.1
pyjwt==2.8.0